+++ package.json
@@ -1,43 +1,37 @@
 {
-  "name": "reverbr8",
+  "name": "discord-chat-bot-revival",
   "version": "1.0.0",
-  "description": "REVerbr8. A bot that helps increase viewer interaction with your video game streams.",
+  "description": "My new Discord Bot",
   "main": "index.js",
-  "dependencies": {
-    "@types/express": "^4.17.8",
-    "@types/mocha": "^8.0.3",
-    "@types/node": "^14.14.2",
-    "chai": "^4.2.0",
-    "discord.js": "^12.3.1",
-    "dotenv": "^8.2.0",
-    "express": "^4.17.1",
-    "helmet": "^4.1.1",
-    "inversify": "^5.0.1",
-    "inversify-express-utils": "^6.3.2",
-    "inversify-logger-middleware": "^3.1.0",
-    "mongodb": "^3.6.3",
-    "reflect-metadata": "^0.1.13",
-    "ts-mockito": "^2.6.1",
-    "ts-node": "^9.0.0",
-    "typescript": "^4.0.3"
-  },
-  "devDependencies": {
-    "mocha": "^8.1.3"
-  },
   "scripts": {
-    "test": "mocha -r ts-node/register \"tests/**/*.spec.ts\"",
     "start": "node src/index.js",
-    "watch": "tsc -p tsconfig.json -w",
-    "comp": "tsc -p tsconfig.json"
+    "build": "tsc -p tsconfig.json -w",
+    "test": "mocha -r ts-node/register \"tests/**/*.spec.ts\""
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/knoxiiw/discord-chat-bot-revival.git"
   },
-  "author": "William Knox",
+  "author": "knoxiiw",
   "license": "ISC",
   "bugs": {
     "url": "https://github.com/knoxiiw/discord-chat-bot-revival/issues"
   },
-  "homepage": "https://github.com/knoxiiw/discord-chat-bot-revival#readme"
+  "homepage": "https://github.com/knoxiiw/discord-chat-bot-revival#readme",
+  "dependencies": {
+    "@types/node": "^16.11.0",
+    "discord.js": "^13.2.0",
+    "dotenv": "^10.0.0",
+    "inversify": "^6.0.1",
+    "reflect-metadata": "^0.1.13",
+    "typescript": "^4.4.4"
+  },
+  "devDependencies": {
+    "@types/chai": "^4.2.22",
+    "@types/mocha": "^9.0.0",
+    "chai": "^4.3.4",
+    "mocha": "^9.1.3",
+    "ts-mockito": "^2.6.1",
+    "ts-node": "^10.3.0"
+  }
 }
