--- src/const/symbols.ts
@@ -1,14 +1,0 @@
-export const SYMBOLS = {
-    // Discord Bot
-    Bot: Symbol("Bot"),
-    Client: Symbol("Client"),
-    Token: Symbol("Token"),
-    MessageResponder: Symbol("MessageResponder"),
-    RobloxQueueService: Symbol("RobloxQueueService"),
-    RobloxGameBuilderService: Symbol("RobloxGameBuilderService"),
-
-    // Express
-    ExpressServer: Symbol("ExpressServer"),
-    MongoDBClient: Symbol("MongoDBClient"),
-    RobloxGameService: Symbol("RobloxGameService")
-};
