--- src/const/tags.ts
@@ -1,3 +1,0 @@
-export const TAGS = {
-    GameQueueController: 'GameQueueController'
-}
