+++ src/index.ts
@@ -1,22 +1,11 @@
 require('dotenv').config();
-import { default as container } from './inversify.config';
-import { SYMBOLS } from './const/symbols';
+import container from './inversify.config';
+import { SYMBOLS } from './symbols';
+import { Bot } from './bot';
 
-import { Bot } from './bot/bot';
-import { ExpressServer } from './express/express-server';
-
-// Discord Bot
 let bot = container.get<Bot>(SYMBOLS.Bot);
-bot.logIn().then(() => {
-    console.log('Logged in');
-    bot.listen();
+bot.listen().then((res) => {
+	console.log('Logged in');
 }).catch((error) => {
-    console.log('error logging in: ' + error)
+	console.log("Error during login: ", error);
 });
-
-// Express
-/*
-let express = container.get<ExpressServer>(SYMBOLS.ExpressServer);
-express.containerLoad();
-express.serverStart();
-*/
