+++ src/symbols.ts
@@ -1,0 +1,5 @@
+export const SYMBOLS = {
+	Bot: Symbol("Bot"),
+	Client: Symbol("Client"),
+	Token: Symbol("Token"),
+}
