+++ tsconfig.json
@@ -1,27 +1,25 @@
 {
-    "compilerOptions" : {
-        "module":"commonjs",
-        "moduleResolution":"node",
-        "target":"ES2018",
-        "lib":[
-            "ES6",
-            "DOM"
-        ],
-        "sourceMap":true,
-        "types":[
-            "node",
-            "reflect-metadata"
-        ],
-        "typeRoots":[
-            "node_modules/@types"
-        ],
-        "experimentalDecorators":true,
-        "emitDecoratorMetadata":true,
-        "forceConsistentCasingInFileNames": true,
-        "resolveJsonModule":true,
-        "removeComments": true
-    },
-    "exclude":[
-        "node_modules"
-    ]
+  "compilerOptions": {
+    "module": "commonjs",
+    "moduleResolution": "node",
+    "target": "ES2016",
+    "lib": [
+      "ES6",
+      "DOM"
+    ],
+    "sourceMap": true,
+    "types": [
+      "node",
+      "reflect-metadata"
+    ],
+    "typeRoots": [
+      "node_modules/@types"
+    ],
+    "experimentalDecorators": true,
+    "emitDecoratorMetadata": true,
+    "resolveJsonModule": true
+  },
+  "exclude": [
+    "node_modules"
+  ]
 }
